overall client refactoring
