+++ space-client/src/app/debug-client/debug-client.component.ts
@@ -3,8 +3,8 @@
  * Created by basst314 on 02.01.2017.
  */
 import { Component, OnInit } from '@angular/core';
-import { SpaceDebugHttpService } from './space-debug-http.service';
-import { SpaceWorld } from '../domain/SpaceWorld';
+import { SpaceDebugHttpService } from './service/space-debug-http.service';
+import { SpaceWorld } from '../domain/space-world';
 
 @Component({
   selector: 'debug-client',
