+++ space-client/src/app/debug-client/service/space-http-api-url.ts
