+++ space-client/src/app/domain/space-world.ts
