+++ space-client/src/app/shared/services/web-socket.service.ts
